% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_animal_unit.R
\name{adult_animal_unit}
\alias{adult_animal_unit}
\title{adult_animal_unit}
\usage{
adult_animal_unit(animal_numbers, coef_tbl, multiplier = NULL)
}
\arguments{
\item{animal_numbers}{number of each animal species.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Calculate adult animal units.
}
\details{
TBD
}
