% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_excretion.R
\name{animal_excretion}
\alias{animal_excretion}
\title{animal excretion}
\usage{
animal_excretion(
  animal_numbers,
  coef_tbl,
  multiplier = NULL,
  type_excretion = "net"
)
}
\arguments{
\item{animal_numbers}{number of each animal species.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}

\item{type_excretion}{gross or net.}
}
\description{
Element content in animal excretion.
}
\details{
TBD
}
