% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_feed.R
\name{animal_feed}
\alias{animal_feed}
\title{animal feed}
\usage{
animal_feed(crop, coef_tbl, multiplier = NULL)
}
\arguments{
\item{crop}{crop biomass as calculated with `crop_production`.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Element content in animal feed.
}
\details{
TBD
}
