% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_feed_consumption.R
\name{animal_feed_consumption}
\alias{animal_feed_consumption}
\title{animal consumption}
\usage{
animal_feed_consumption(animal_numbers, coef_tbl, multiplier = NULL)
}
\arguments{
\item{animal_numbers}{number of each animal species.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Quantity of an element ingested by animals.
}
\details{
Animal consumption is calculated according tothe following equation:

\deqn{afc=animal\cdot consumption_rate}

where afc is the animal feed consumption, animal is the the number of each animal species, consumption_rate is the nutrient consumption.
For example, nitrogen has a consumption_rate usually expressed as kg ind^-1^ year^-1^.
}
