% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_numbers.R
\docType{data}
\name{animal_numbers}
\alias{animal_numbers}
\title{Prices of 50,000 round cut diamonds.}
\format{
An object of class \code{data.frame} with 38 rows and 37 columns.
}
\usage{
animal_numbers
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
