% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_production.R
\name{animal_production}
\alias{animal_production}
\title{animal production}
\usage{
animal_production(animal_numbers, coef_tbl, multiplier = NULL)
}
\arguments{
\item{animal_numbers}{number of each animal species.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Animal production for human consumption.
}
\details{
TBD
}
