% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nani.R
\name{as_nani}
\alias{as_nani}
\title{as_nani}
\usage{
as_nani(x, input, multiplier = NULL)
}
\arguments{
\item{x}{data to transform into the `nani` format.}

\item{input}{the type of input.}

\item{multiplier}{scaling factor (e.g. to calculate hectars instead of km^2^).}
}
\description{
Returns a tibble that fit with the `nani` package format.
}
\details{
Inputs can be calculated in several ways dependending on the available datasets.
The `nani` package implements widely used calculations methods, but to cover the entire set of possibile calculation methods is unrealistic.
The function `as_nani` allows to transfrom user data into a format suitable to the `nani` package. For instance,
data on animal excretion can be calculated with a different method, but transformed into a format suitable to the `nani` package.
Several options are provided, use the function `as_nani_options` for details.
}
\seealso{
as_nani_options
}
