% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_nani_options.R
\name{as_nani_options}
\alias{as_nani_options}
\title{as_nani options}
\usage{
as_nani_options(element = "all")
}
\arguments{
\item{element}{can be `all`, `nitrogen` and `phosphorus`.}
}
\description{
Options to import datasets into nani format.
}
\details{
Data can be obtained with algorithms not implemented in the `nani` an imported into the `nani`
package format with `as_nani`. The function `as_nani_options` provides codes to import data for the target element.
}
\examples{

# list all of the codes for nitrogen
as_nani_options(element = "nitrogen")
}
\seealso{
as_nani
}
