% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_table.R
\docType{data}
\name{coef_table}
\alias{coef_table}
\title{Prices of 50,000 round cut diamonds.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 825 rows and 5 columns.
}
\usage{
coef_table
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
