% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\docType{data}
\name{crop}
\alias{crop}
\title{Prices of 50,000 round cut diamonds.}
\format{
An object of class \code{data.frame} with 38 rows and 91 columns.
}
\usage{
crop
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
