% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_production.R
\name{crop_production}
\alias{crop_production}
\title{calculate nitrogen from crops}
\usage{
crop_production(crop, coef_tbl, multiplier = NULL)
}
\arguments{
\item{crop}{crop for each species.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Calculate nitrogen from cultivation.
}
\details{
TBD
}
