% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizers.R
\name{fertilizers}
\alias{fertilizers}
\title{fertilizers}
\usage{
fertilizers(fertilizer_quantity, coef_tbl, multiplier = NULL)
}
\arguments{
\item{fertilizer_quantity}{fertilizers application.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Fertilizers application.
}
\details{
TBD
}
