% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_excretion.R
\name{human_excretion}
\alias{human_excretion}
\title{human excretion}
\usage{
human_excretion(human, coef_tbl, multiplier = 0.001)
}
\arguments{
\item{human}{number of inhabitants.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Element content in human excretion.
}
\details{
TBD
}
