% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_food.R
\name{human_food}
\alias{human_food}
\title{human food}
\usage{
human_food(crop, coef_tbl, multiplier = NULL)
}
\arguments{
\item{crop}{crop biomass as calculated with `crop_production`.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Element content in human food.
}
\details{
TBD
}
