% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_food_consumption.R
\name{human_food_consumption}
\alias{human_food_consumption}
\title{human consumption}
\usage{
human_food_consumption(human, coef_tbl, multiplier = NULL)
}
\arguments{
\item{human}{number of inhabitants.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Quantity of an element ingested by humans.
}
\details{
TBD
}
