% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_non_food_production.R
\name{human_non_food_production}
\alias{human_non_food_production}
\title{human non food production}
\usage{
human_non_food_production(human, coef_tbl, multiplier = NULL)
}
\arguments{
\item{human}{number of inhabitants.}

\item{coef_tbl}{coefficients.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
Quantity of an element produced by humans for factors other than food.
}
\details{
TBD
}
