% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_anthropogenic_input.R
\name{net_anthropogenic_input}
\alias{net_anthropogenic_input}
\title{net anthropogenic input}
\usage{
net_anthropogenic_input(..., element = "nitrogen")
}
\arguments{
\item{...}{objects needed for calculated anthropogenic input (see details).}

\item{element}{can be one of `nitrogen` and `phosphorus`.}
}
\description{
Calculates the net anthropogenic input of the desired variable.
}
\details{
TBD
}
