% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_fixation.R
\name{nitrogen_fixation}
\alias{nitrogen_fixation}
\title{nitrogen fixation}
\usage{
nitrogen_fixation(cultivated_area, crop, coef_tbl, fix_tbl, multiplier = NULL)
}
\arguments{
\item{cultivated_area}{crop area for each species.}

\item{crop}{crop biomass for each species.}

\item{coef_tbl}{coefficients.}

\item{fix_tbl}{data.frame with a list of nitrogen fixing species as specified below.}

\item{multiplier}{scaling factor (e.g. to calculate hectars instead of km^2^). It can be a vector to multiplier
different database.}
}
\description{
Symbiotic and non-symbiotic nitrogen fixation.
}
\details{
TBD
}
