% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.R
\name{plot_maps}
\alias{plot_maps}
\title{plot maps}
\usage{
plot_maps(
  x,
  shape,
  sp_unit = NULL,
  variable = NULL,
  alpha = 1,
  background = NULL
)
}
\arguments{
\item{x}{a data.frame.}

\item{shape}{a shapefile.}

\item{sp_unit}{spatial unit to plot.}

\item{variable}{variable to plot}

\item{alpha}{set transparency.}

\item{background}{set an OpenStreetMaps backround. Default to `NULL`.
Options are many, useful are "open-street-map" and "stamen-terrain".
See plotly reference \href{https://plotly.com/r/reference/layout/mapbox/}{here} in the style section.}
}
\description{
plot maps
}
\details{
Yield is calculated as the product between cultivated area and biomass
}
