% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_yield.R
\name{plot_yield}
\alias{plot_yield}
\title{plot yield}
\usage{
plot_yield(yield, relative = TRUE, add_bound = TRUE)
}
\arguments{
\item{yield}{a yield object.}

\item{relative}{scale all values to unit.}

\item{add_bound}{add boundaries to level 1 properties.}
}
\description{
plot yield
}
\details{
Yield is calculated as the product between cultivated area and biomass
}
