% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_system_budget.R
\name{soil_system_budget}
\alias{soil_system_budget}
\title{net input}
\usage{
soil_system_budget(..., element = "nitrogen")
}
\arguments{
\item{...}{objects needed for calculated anthropogenic input (see details).}

\item{element}{can be one of `nitrogen` and `phosphorus`.}
}
\description{
Calculates the net anthropogenic input of the desired variable.
}
\details{
TBD
}
