% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed_percentage.R
\name{watershed_percentage}
\alias{watershed_percentage}
\title{watershed percentage}
\usage{
watershed_percentage(x, percentage, scale = NULL)
}
\arguments{
\item{x}{a `data.frame` or `tibble`.}

\item{percentage}{a `data.frame` or `tibble` containing the percentage of each spatial unit in a watershed.}

\item{scale}{scaling factor (e.g. to calculate hectars instead of km^2^).}
}
\description{
Normalise data to watershed area in each spatial unit.
}
\details{
Administrative boundaries often does not match with watershed boundaries. The function `watershed_percentage`
normalise data to watershed area in each spatial unit. The x argument need a column called "sp_unit" and another
column with percentages of each spatial unit in a watershed (0-1). The second can have any name.
}
