% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{yield}
\alias{yield}
\title{calculate yield}
\usage{
yield(cultivated, crop, multiplier = NULL)
}
\arguments{
\item{cultivated}{cultivated area for each species.}

\item{crop}{crop for each species.}

\item{multiplier}{multiply by a number (e.g. to calculate hectars instead of km^2^).}
}
\description{
calculate yield from total crop and cultivated area
}
\details{
Yield is calculated as the product between cultivated area and crop
}
